+++ ficary/sites.py
@@ -33,6 +33,7 @@ from .erotica import (
 from .ficwad import FicWadScraper
 from .mediaminer import MediaMinerScraper
 from .royalroad import RoyalRoadScraper
+from .scribblehub import ScribbleHubScraper
 from .scraper import BaseScraper, FFNScraper
 from .scraper import ensure_scheme as _ensure_scheme
 from .wattpad import WattpadScraper
@@ -234,6 +235,7 @@ _HOSTNAME_TO_SCRAPER: list[tuple[str, type[BaseScraper]]] = [
     ("archiveofourown.org", AO3Scraper),
     ("ao3.org", AO3Scraper),
     ("royalroad.com", RoyalRoadScraper),
+    ("scribblehub.com", ScribbleHubScraper),
     ("mediaminer.org", MediaMinerScraper),
     ("literotica.com", LiteroticaScraper),
     ("wattpad.com", WattpadScraper),
@@ -266,6 +268,7 @@ ALL_SCRAPERS: list[type[BaseScraper]] = [
     FicWadScraper,
     AO3Scraper,
     RoyalRoadScraper,
+    ScribbleHubScraper,
     MediaMinerScraper,
     LiteroticaScraper,
     WattpadScraper,